pyproject.toml
[project]
name = "video-translator"
version = "0.1.0"
description = "Automatically translate English videos to Hindi using OpenAI APIs (Whisper, GPT-4, TTS) and FFmpeg"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Video Translator Project"},
]
keywords = ["video", "translation", "openai", "whisper", "tts", "ffmpeg", "hindi", "english"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]
dependencies = [
    "ffmpeg-python>=0.2.0",
    "openai>=1.93.0",
    "python-dotenv>=1.1.1",
    "yt-dlp>=2025.6.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.scripts]
video-translator = "main:main"
check-requirements = "check_requirements:main"

[project.urls]
Homepage = "https://github.com/your-username/video-translator"
Repository = "https://github.com/your-username/video-translator"
Issues = "https://github.com/your-username/video-translator/issues"
Documentation = "https://github.com/your-username/video-translator#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
include = [
    "main.py",
    "check_requirements.py",
    "example_usage.py",
    "env.template"
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]
